+++ NEWS.md
@@ -5,6 +5,8 @@
 * Use unicode em-dash and no spaces in `lbl_dash()` and `lbl_discrete()`.
 * The `groups` argument to `chop_evenly()`,  deprecated in 0.4.0, has 
   been removed.
+* Arguments `first` and `last` can now be used in `lbl_intervals()` to override
+  the first and last interval labels.
 
 # santoku 0.6.0